requirements/test.txt
-r base.txt

django-markup>=1.2
django-nose>=1.4,<1.5
coverage<3.8
factory-boy<2.6
mock<1.1
textile>=2.3.7
tox<3
